Added ImGui and cook-torrance brdf lighting in fragment shader

+++ Gaia/src/Gaia/Material.h
@@ -7,8 +7,8 @@ namespace Gaia
 struct Material
 {
 	glm::vec4 baseColorFactor = glm::vec4(1.0);
-	int metallicFactor = 0.0;
-	int roughnessFactor = 1.0;
+	float metallicFactor = 0.0;
+	float roughnessFactor = 1.0;
 	float normalStrength = 1.0;
 
 	int baseColorTexture = -1;
